+++ src/main/kotlin/invest/collect/com/routes/Routing.kt
@@ -2,11 +2,9 @@ package invest.collect.com.routes
 
 import invest.collect.com.controllers.FinancialController
 import invest.collect.com.entities.Message
-import invest.collect.com.entities.Transaction
 import invest.collect.com.entities.Wallet
 import io.ktor.http.*
 import io.ktor.server.application.*
-import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import kotlinx.serialization.encodeToString
@@ -20,7 +18,7 @@ fun Application.configureRouting() {
     routing {
         put("/topUp/{userId}/{amount}") {
             val userId: Long = call.parameters["userId"]!!.toLong()
-            val amount: Int = call.parameters["amount"]!!.toInt()
+            val amount: Double = call.parameters["amount"]!!.toDouble()
             try {
                 walletController.topUp(url, userId, amount)
                 call.respond(HttpStatusCode.OK, "$amount")
@@ -35,14 +33,14 @@ fun Application.configureRouting() {
 
         post("/buy/{userId}/{amount}") {
             val userId: Long = call.parameters["userId"]!!.toLong()
-            val amount: Int = call.parameters["amount"]!!.toInt()
+            val amount: Double = call.parameters["amount"]!!.toDouble()
             val status = walletController.buyCollectible(url, userId, amount)
             call.respond(status)
         }
 
         post("/sell/{userId}/{amount}") {
             val userId: Long = call.parameters["userId"]!!.toLong()
-            val amount: Int = call.parameters["amount"]!!.toInt()
+            val amount: Double = call.parameters["amount"]!!.toDouble()
             val status = walletController.sellCollectible(url, userId, amount)
             call.respond(status)
         }
